server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fastapi>=0.109.2",
    "groq>=0.22.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "numpy==2.2.4",
    "pydantic>=2.6.1",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.1.0",
    "python-dotenv>=1.0.1",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "uvicorn>=0.27.1",
]
